Actionable Insights and Recommendations
The top 5 countries with the highest spending statistics are China, South Africa, Switzerland, United Kingdom and United State of America with United Kingdom having the highest spending tourists (50%).

Tourists within the age of 45-64 are the highest spenders and the overall highest spenders are tourists within the age of 45-64 who travelled with their spouses. The highest age group among the tourists are mostly young people between the age of 25-44. Since the most populated tourists are young people who are about 50% of the general population, more attention should be given to what interest this age group. for example, more investment should be made in the activities that is of interest to this age group.

The analysis also shows that the top main activity of tourists is wildlife tourism and this is the aspect of tourism that is more profitable and therefore worthwhile to invest in. more area should be allocated for wildlife.

Also from the analysis, tourists who do not consume package food are more than tourists who do. Indicating that the tourist mostly come alongside with food from their countries. It could be that the tourists are not familiar with the Tanzanian's kind of food. The Tanzanians should consider the production of diverse kinds of food that the tourists are familiar with. This will lead to an increase in the consumption of package food.